Coursera applying for Financial Aid
## Why are you applying for Financial Aid? 

I’m a student from Bangladesh and want to learn Big Data Essentials. I think it will be beneficial for me to get into a good firm as an intern. But I’ve no job of my own to carry the expanses to pay for the certificate of this course. I live only for my scholarship. In this circumstance, it is very much difficult for me to gather such amount of money for the certificate. Financial Aid will help me take this course without any adverse impact on my monthly essential needs. So, I’m badly in need of this financial aid. Receiving this Financial Aid will open for me a new horizon of the world of Coursera courses, which in turn will help me in the future. Sir, we are four brothers and we all are at the learning stage and it becomes a lot difficult for the family members to pay the whole amount for all the three of us and if I add up my course it will make even harder for them to pay. Sir, I need this course very badly for my CV and for increasing my knowledge about the subject.


## How will taking this course help you achieve your career goals?

I want to take this course as I want to learn Big Data Essentials. I want to complete the course due to my curiosity and also that I can put a good CV to get applied to a job. This course will boost my job prospects after graduation from my institute. It will help perform better in carrying out various programs in a computer language and give me an edge over my competitors. A verified certificate will attach credibility to the certificate I receive from this course. I plan to complete all assignments on or before time as I have done in previous Signature Track Courses. Also, I intend to participate in Discussion Forums, which I have found to supplement my learning immensely in the other online courses I have taken on Coursera. I also plan to grade assignments that are to peer-reviewed which I believe will an invaluable learning opportunity.

## Would you consider using a low-interest loan to pay for your courses? No then Explain

Sir, the financial status of the family is not too good to pay the loan amount. We already have a lot of dept in the bank and my parents are paying it on a regular basis. It would make their life even harder to add a new loan over them. Sir, I don't want to put any pressure over them. Sir, it would be a great help for me to get a good job and help my family if I'm able to get this course.